requirements.txt
musdb==0.4.0
tensorboard==2.13.0
git+https://github.com/KinWaiCheuk/AudioLoader.git
hydra-core==1.3.2
lightning-utilities==0.8.0
pytorch-lightning==2.0.2
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.0.0+cu118
torchvision==0.15.1+cu118
torchaudio==2.0.1

logo
